[config] add installer for some vim plugins
I do not want to add submodules, since these references can break any
time. Also, using submodules would cause the check script to check for a
lot of files which is not nescessary.
I do not considder the exact internal state of these plugins part of my
shared configuration, so I do not need to track it.