[tools] rework for screencast script
Pulseaudio is very flaky for this kind of setup, there are often
situations where AV goes horribly out of sync. If a device goes
idle there are a lot of warnings. There needs to be
some way to either do SRC directly or capture streams separately
and resample and merge afterwards manually.